+++ components/chat/messages/index.tsx
@@ -1,8 +1,8 @@
-import { UserMessage } from './user-message';
-import { AssistantMessage } from './assistant-message';
-import { FileAttachmentDisplay } from './file-attachment-display';
-import { UserMessageActions } from './user-message-actions';
-import { AssistantMessageActions } from './assistant-message-actions';
+import { UserMessage } from './user-message/user-message';
+import { AssistantMessage } from './user-message/assistant-message';
+import { FileAttachmentDisplay } from './user-message/file-attachment-display';
+import { UserMessageActions } from './user-message/user-message-actions';
+import { AssistantMessageActions } from './user-message/assistant-message-actions';
 
 export { 
   UserMessage,
