+++ components/chat/messages/user-message/file-attachment-display.tsx
